# src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "RVM API",
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/organisations": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Return all organisations",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Create organisation",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Organisation name.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "website",
                        "in": "query",
                        "description": "Organisation website.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contact_person",
                        "in": "query",
                        "description": "Organisation Contact Person.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Organisation email.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Organisation phone.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "county",
                        "in": "query",
                        "description": "Organisation county.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Organisation city.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Organisation address.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "comments",
                        "in": "query",
                        "description": "Organisation comments.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/organisations/{id}": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Show organisation info ",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not acceptable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Update organisation",
                "operationId": "update",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Delete organisation",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/organisations/{id}/email": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Send notification via email to a organisation admin",
                "operationId": "send",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/organisations/{id}/volunteers": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Show all volunteers of an Organisation ",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/organisations/{id}/resources": {
            "get": {
                "tags": [
                    "Organisations"
                ],
                "summary": "Show all resources of an Organisation ",
                "operationId": "show",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
...
}